+++ src/frontend/apps/impress/src/core/conf.ts
@@ -6,10 +6,10 @@ export const baseApiUrl = (apiVersion: string = '1.0') => {
   return `${origin}/api/v${apiVersion}/`;
 };
 
-export const signalingUrl = (padId: string) => {
+export const signalingUrl = (docId: string) => {
   const base =
     process.env.NEXT_PUBLIC_SIGNALING_URL ||
     (typeof window !== 'undefined' ? `wss://${window.location.host}/ws` : '');
 
-  return `${base}/${padId}`;
+  return `${base}/${docId}`;
 };