+++ src/frontend/apps/impress/src/features/docs/doc-management/components/index.ts
@@ -1,0 +1,3 @@
+export * from './CardCreateDoc';
+export * from './ModalRemoveDoc';
+export * from './ModalUpdateDoc';