+++ src/frontend/apps/impress/src/features/pads/pad-management/components/index.ts
@@ -1,3 +1,0 @@
-export * from './CardCreatePad';
-export * from './ModalRemovePad';
-export * from './ModalUpdatePad';