+++ CHANGELOG.md
@@ -32,6 +32,8 @@ and this project adheres to
 - ⚡️(e2e) unique login between tests (#80)
 - ⚡️(CI) improve e2e job (#86)
 - ♻️(frontend) improve the error and message info ui (#93)
+- ♻️(frontend) improve the error and message info ui (#93)
+- ✏️(frontend) change all occurences of pad to doc (#99)
 
 ## Fixed